+++ doc/CHANGELOG.md
@@ -21,6 +21,7 @@
 - Support for D6 Legends
 - Support for World of Darkness Homebrew 10s cancel 1s
 - Support for Vampire 5e
+- Support for Laser and Feelings
 - "stability and performance improvements"
 
 ## [1.3.0] - 2025-07-03